backend/pyproject.toml
[tool.poetry]
name = "applytrack"
version = "0.1.0"
description = "A personal 'mini-ATS' to track applications."
authors = ["Jules <jules@example.com>"]
packages = [{include = "applytrack", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
celery = "^5.4.0"
redis = "^5.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
openai = "^1.30.0"
python-multipart = "^0.0.9"
argon2-cffi = "^25.1.0"
nest-asyncio = "^1.6.0"

[tool.poetry.dev-dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
